// apps/web/generated/api/pokemon.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * Poke - Fight
 * OpenAPI spec version: 1.0
 */

export interface Pokemon {
  /** Pokémon description */
  description: string;
  /** Unique Pokémon identifier */
  id: string;
  /** Pokémon image URL */
  image: string;
  /** Pokémon name */
  name: string;
  /** Pokémon id */
  pokemonId: number;
  /** Pokémon types */
  types: string[];
}