// apps/web/generated/api/statistic.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * Poke - Fight
 * OpenAPI spec version: 1.0
 */

export interface Statistic {
  /** Unique Statistic identifier */
  id: string;
  /** Dislikes */
  pass: number;
  /** Pokemon id */
  pokemonId: number;
  /** Likes */
  smash: number;
}